pyproject.toml
[project]
name = "ushareiplay"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.19.0",
    "appium-python-client==3.1.0",
    "langdetect>=1.0.9",
    "lxml>=5.0.0",
    "pillow>=12.1.1",
    "pytesseract==0.3.10",
    "pytest>=7.4.3",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.0.0",
    "pyyaml==6.0.1",
    "selenium==4.15.2",
    "tortoise-orm>=0.19.0",
    "urllib3==2.0.7",
]

[project.scripts]
ushareiplay = "ushareiplay.__main__:run"

[tool.pytest.ini_options]
asyncio_mode = "auto"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
